Fix failing target group updates
Target group updates were failing with the following message: "target group is currently in use by a listener or a rule"

Using create_before_destroy is a workarounf for this issue.

As a consequence, we also need to remove the name for the target groups. Otherwise, the creation fails because the name already exists.