Add the migration script
(this script will be required to seamlessly upgrade to the next version of the load-balancer module, >=1.7.0)

Context: in terraform, the egress/ingress rules for a security group can be declared "inline" (inside the terraform "aws_security_group" resource) or as individual "aws_security_group_rule" resources.

Unfortunately, when Terraform tries to update an existing infrastructure to migrate from "inline" to separate rules, it does not know that the separate rules are the same as the inline rules. Therefore, Terraform will try to create new rules, but there will be a conflict in AWS.

Instead, this script imports the existing AWS rules declared "inline" in the security groups, into the terraform state as separate "aws_security_group_rule" resources.