Fix resources creation issue with for_each
We had two issues with the previous commit which introduced using
for_each to avoid recreating all ASGs when a single ASG was added or
deleted : the target_group_arn and the autoscaling_group_name are unknown
during apply thus they cannot be used as a key of the for_each.

The solution is to use the port index instead of the target_group_arn, and
the autoscaling group key (user provided) instead of the autoscaling group
name (generated by AWS).

This will unfortunately re-break the state and heavily complexify the
migration script.